Extract Launchplane preview index renderer
